Date Model

S4 classes to store the event and accumulation times of archaeological assemblages.

Slots

data

A numeric matrix of count data.

dates

A numeric vector of dates.

model

A multiple linear model: the Gaussian multiple linear regression model fitted for event date estimation and prediction.

cutoff

An integer vector giving the cutoff value.

dimension

An integer vector giving the CA dimensions kept.

level

A length-one numeric vector giving the confidence level.

row_events

A four columns numeric matrix giving the predicted event dates for each archaeological assemblage, with the following columns:

date

The event date estimation.

lower

The lower boundary of the confidence interval.

upper

The upper boundary of the confidence interval.

error

The standard error of predicted dates.

column_events

A four columns numeric matrix giving the predicted event dates for each archaeological type or fabric, with the following columns:

date

The event date estimation.

lower

The lower boundary of the confidence interval.

upper

The upper boundary of the confidence interval.

error

The standard error of predicted dates.

accumulation

A two columns numeric matrix giving the point estimate of the accumulation dates and the corresponding error.

Subset

In the code snippets below, x is a DateEvent object.

x[[i]]

Extracts information from a slot selected by subscript i. i is a length-one character vector. Returns the corresponding slot values.

See also

Author

N. Frerebeau